profiles: introduce switch profile
This adds a generic switch package profile, it provides
common tools that are used by all switches.

Its aimed to reduce code duplication as more switch
devices are added.

While we are here, drop nftables and tcpdump.
Those are firewall and debugging tools which are not
commonly used on switches.

Signed-off-by: Robert Marko <robert.marko@sartura.hr>